ImagesTable: Split `Status` to individual component files
This splits the original `Status` file to individual component files for better readability.
