+++ src/Components/ImagesTable/components/Status/index.tsx
@@ -1,0 +1,9 @@
+export { default as AwsDetailsStatus } from './components/AwsDetailsStatus';
+export { default as Azurestatus } from './components/Azurestatus';
+export { default as CloudStatus } from './components/CloudStatus';
+export { default as ErrorStatus } from './components/ErrorStatus';
+export { default as ExpiringStatus } from './components/ExpiringStatus';
+export { default as LocalStatus } from './components/LocalStatus';
+export { default as ProgressStatus } from './components/ProgressStatus';
+export { default as Status } from './components/Status';
+export * from './statusConfig';
